Prepare to replace Project columns with reference to Idea (#3)
## What did we change?

* Added a reference from the `Project` model to the `Idea` model
* Ignored all the columns on `Project` that are also on `Idea`
* Added some basic model specs
* Moved feature specs into `spec/features`

## Why are we doing this?

The spec changes were in support of the other changes, just so there were some specs to build against.

The change to `Project` removes the need to copy info between these two models. This isn't finished yet - in a future PR we'll drop the columns on `projects` that we no longer need and add the UI to build projects. However, we need to release the `ignored_columns` change first, so this PR has to go out before we can make any further changes.