Enable voting on projects (#15)
## What did we change?

Added the ability for a user to vote on all the awards at a particular event, voting for one award at a time.

This keeps the state (who's voting, where they are in the list of awards, whether they've finished voting) in URL params. This isn't particularly reliable - users could easily get around this and vote multiple times - but we can't really deal with that well until we have a `User` model and authentication.

## Why are we doing this?

So people can vote on which projects deserve awards at our current hackathon.