Fix bug causing autorefresh after navigation (#33)
## What did we change?

Implemented autorefresh of awards page in JS just on the awards page rather than in application layout.

## Why are we doing this?

When users navigated away from the awards page, the autorefresh was still happening, potentially because turbolinks doesn't replace the `<head>`. 

This uses JS and only reloads the page if it is the same URL as at the start of the timeout to make sure the user has not navigated away.